+++ include/sentry.h
@@ -1363,6 +1363,16 @@ SENTRY_API sentry_user_consent_t sentry_user_consent_get(void);
 SENTRY_API sentry_uuid_t sentry_capture_event(sentry_value_t event);
 
 /**
+ * Allows capturing independently created minidumps.
+ *
+ * This generates a fatal error event, includes the scope and attachments.
+ * If the event isn't dropped by a before-send hook, the minidump is attached
+ * and the event is sent.
+ */
+SENTRY_API void sentry_capture_minidump(const char *path);
+SENTRY_API void sentry_capture_minidump_n(const char *path, size_t path_len);
+
+/**
  * Captures an exception to be handled by the backend.
  *
  * This is safe to be called from a crashing thread and may not return.